[Clean] LikeButtonコンポーネントを別ファイルに切り出し
